pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "line-bot-sdk>=3.17.1",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
